requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-whois==0.8.0
dnspython==2.4.2
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiomysql==0.2.0
greenlet==3.0.2

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# 限流
slowapi==0.1.9

# 环境变量
python-dotenv==1.0.0
